// features/api/model/customersIndex200.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */
import type { CustomerResource } from './customerResource';
import type { CustomersIndex200Links } from './customersIndex200Links';
import type { CustomersIndex200Meta } from './customersIndex200Meta';

export type CustomersIndex200 = {
  data: CustomerResource[];
  links: CustomersIndex200Links;
  meta: CustomersIndex200Meta;
};