// features/api/model/storeCustomerRequest.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */

export interface StoreCustomerRequest {
  /** @maxLength 255 */
  name: string;
  email: string;
  /**
   * @maxLength 20
   * @nullable
   */
  phone?: string | null;
  /**
   * @maxLength 50
   * @nullable
   */
  tax_pin?: string | null;
  /**
   * @minimum 0
   * @nullable
   */
  credit_limit?: number | null;
}