// features/api/model/customersIndex200Meta.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */
import type { CustomersIndex200MetaLinksItem } from './customersIndex200MetaLinksItem';

export type CustomersIndex200Meta = {
  /** @minimum 1 */
  current_page: number;
  /**
   * @minimum 1
   * @nullable
   */
  from: number | null;
  /** @minimum 1 */
  last_page: number;
  /** Generated paginator links. */
  links: CustomersIndex200MetaLinksItem[];
  /**
   * Base path for paginator generated URLs.
   * @nullable
   */
  path: string | null;
  /**
   * Number of items shown per page.
   * @minimum 0
   */
  per_page: number;
  /**
   * Number of the last item in the slice.
   * @minimum 1
   * @nullable
   */
  to: number | null;
  /**
   * Total number of items being paginated.
   * @minimum 0
   */
  total: number;
};