// features/api/model/customersIndex200Links.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */

export type CustomersIndex200Links = {
  /** @nullable */
  first: string | null;
  /** @nullable */
  last: string | null;
  /** @nullable */
  prev: string | null;
  /** @nullable */
  next: string | null;
};