// features/api/model/storeEditRequestRequest.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */

export interface StoreEditRequestRequest {
  model_type: string;
  model_id: string;
  requested_data: string[];
  /** @nullable */
  reason?: string | null;
}